Restore checking github for youtube-dl versions
praise the eff!

+++ Classes/UpdateChecker.cs
@@ -15,6 +15,7 @@ namespace youtube_dl_gui {
         public static bool bypassDebug = false;
         public static GitData GitData = GitData.GetInstance();
         public static Verification verif = Verification.GetInstance();
+        private static bool DMCA = false; // Will bypass the youtube-dl check if it gets DMCA'd
 
         public static void CheckForUpdate(bool ForceCheck = false) {
             if (Program.IsDebug && !ForceCheck) {
@@ -79,7 +80,9 @@ namespace youtube_dl_gui {
         }
 
         public static bool CheckForYoutubeDlUpdate() {
-            return false; // repo was dcma'd, can't check for updates.
+            if (DMCA) {
+                return false;
+            }
 
             if (GitData.YoutubeDlUpdateAvailable) {
                 return GitData.YoutubeDlUpdateAvailable;
@@ -108,23 +111,26 @@ namespace youtube_dl_gui {
             Environment.Exit(0);
         }
         public static void UpdateYoutubeDl() {
-            using (WebClient wc = new WebClient()) {
-                ServicePointManager.SecurityProtocol = SecurityProtocolType.Tls12;
-                wc.Headers.Add(HttpRequestHeader.UserAgent, "youtube-dl-gui/" + Properties.Settings.Default.appVersion);
+            if (DMCA) {
+                using (WebClient wc = new WebClient()) {
+                    ServicePointManager.SecurityProtocol = SecurityProtocolType.Tls12;
+                    wc.Headers.Add(HttpRequestHeader.UserAgent, "youtube-dl-gui/" + Properties.Settings.Default.appVersion);
 
-                if (File.Exists(verif.YoutubeDlPath)) {
-                    if (File.Exists(verif.YoutubeDlPath + ".old")) {
-                        File.Delete(verif.YoutubeDlPath + ".old");
+                    if (File.Exists(verif.YoutubeDlPath)) {
+                        if (File.Exists(verif.YoutubeDlPath + ".old")) {
+                            File.Delete(verif.YoutubeDlPath + ".old");
+                        }
+                        File.Move(verif.YoutubeDlPath, verif.YoutubeDlPath + ".old");
                     }
-                    File.Move(verif.YoutubeDlPath, verif.YoutubeDlPath + ".old");
+
+                    if (verif.YoutubeDlPath == null) {
+                        verif.SetYoutubeDLPath = Environment.CurrentDirectory + "\\youtube-dl.exe";
+                    }
+                    wc.DownloadFile("https://yt-dl.org/downloads/latest/youtube-dl.exe", verif.YoutubeDlPath);
                 }
 
-                if (verif.YoutubeDlPath == null) {
-                    verif.SetYoutubeDLPath = Environment.CurrentDirectory + "\\youtube-dl.exe";
-                }
-                wc.DownloadFile("https://yt-dl.org/downloads/latest/youtube-dl.exe", verif.YoutubeDlPath);
+                return;
             }
-            return;
 
             GetGitVersionString(1);
